rename ui components so they don't confuse stubborn oldies
